requirements.txt
pandas
numpy
streamlit
nltk
python-docx
requests
bs4
regex
dill
openpyxl
-e .